requirements.txt
pymongo
jsonrpclib-pelix
